Move fail-fast auth tests into CI-collected tests/airflow tree
CI runs 'pytest tests' (python_tests.yml), which only collects the repo-root
tests/ tree, so tests under airflow/astro/tests are never run in CI. Move the
new connection tests into the existing tests/airflow/test_databricks_utils.py
(its conftest already puts airflow/astro on sys.path), and drop the redundant
sys.path / module-stub block since the root tree imports the real packages.
